+++ CHANGELOG.md
@@ -36,6 +36,9 @@ You can find our backwards-compatibility policy [here](https://github.com/hynek/
 - Same for *columns*.
   [#757](https://github.com/hynek/structlog/pull/757)
 
+- Same for *colors* and *force_colors*.
+  [#759](https://github.com/hynek/structlog/pull/759)
+
 - Added `structlog.dev.ConsoleRenderer.get_default_column_styles` for reuse the default column styles.
   [#741](https://github.com/hynek/structlog/pull/741)